refactor layer into own source and header
